perf: Subset nonce should be uint256

+++ test/foundry/utils/ProtocolHelpers.sol
@@ -19,7 +19,7 @@ contract ProtocolHelpers is TestHelpers, TestParameters {
 
     function _createSingleItemMakerAskOrder(
         uint112 askNonce,
-        uint112 subsetNonce,
+        uint256 subsetNonce,
         uint256 strategyId,
         uint256 assetType,
         uint256 orderNonce,
@@ -55,7 +55,7 @@ contract ProtocolHelpers is TestHelpers, TestParameters {
 
     function _createMultiItemMakerAskOrder(
         uint112 askNonce,
-        uint112 subsetNonce,
+        uint256 subsetNonce,
         uint256 strategyId,
         uint256 assetType,
         uint256 orderNonce,
@@ -87,7 +87,7 @@ contract ProtocolHelpers is TestHelpers, TestParameters {
 
     function _createSingleItemMakerBidOrder(
         uint112 bidNonce,
-        uint112 subsetNonce,
+        uint256 subsetNonce,
         uint256 strategyId,
         uint256 assetType,
         uint256 orderNonce,
@@ -122,7 +122,7 @@ contract ProtocolHelpers is TestHelpers, TestParameters {
 
     function _createMultiItemMakerBidOrder(
         uint112 bidNonce,
-        uint112 subsetNonce,
+        uint256 subsetNonce,
         uint256 strategyId,
         uint256 assetType,
         uint256 orderNonce,
